\name{MIXFIM-package}
\alias{MIXFIM-package}
\alias{MIXFIM}
\docType{package}
\title{
Evaluation of the FIM in NLMEMs using MCMC
}
\description{Evaluation and optimization of the Fisher Information Matrix in NonLinear Mixed Effect Models using Markov Chains Monte Carlo for continuous and discrete data. \cr
}
\author{
Marie-Karelle Riviere-Jourdan <eldamjh@gmail.com> and France Mentre <france.mentre@inserm.fr>
}
\references{
Riviere, M-K., Ueckert, S. and Mentre, F,. Evaluation of the Fisher information matrix in nonlinear mixed effect models using Markov Chains Monte Carlo.
}
\details{
\tabular{ll}{
Package: \tab MIXFIM\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2018-11-16\cr
License: \tab GPL-3\cr
}
}
